refactor(configs): move validate() onto OzmuxConfigs
